@@ -2,7 +2,7 @@
  * Copyright (c) The JETSCAPE Collaboration, 2018
  *
  * Modular, task-based framework for simulating all aspects of heavy-ion collisions
- * 
+ *
  * For the list of contributors see AUTHORS.
  *
  * Report issues at https://github.com/JETSCAPE/JETSCAPE/issues
@@ -95,6 +95,23 @@ void Hadronization::WriteTask(weak_ptr<JetScapeWriter> w) {
   } else {
     f->WriteComment("There are no Hadrons");
   }
+}
+
+void Hadronization::DeleteHadrons() {
+  outHadrons.clear();
+}
+
+void Hadronization::DeleteRealHadrons() {
+  outHadrons.erase(
+    std::remove_if(
+      outHadrons.begin(),
+      outHadrons.end(),
+      [](const std::shared_ptr<Hadron>& hadron) {
+          return hadron->pstat() > 0;
+      }
+    ),
+    outHadrons.end()
+  );
 }
 
 } // namespace Jetscape
