@@ -2,7 +2,7 @@
  * Copyright (c) The JETSCAPE Collaboration, 2018
  *
  * Modular, task-based framework for simulating all aspects of heavy-ion collisions
- * 
+ *
  * For the list of contributors see AUTHORS.
  *
  * Report issues at https://github.com/JETSCAPE/JETSCAPE/issues
@@ -139,7 +139,7 @@ void HadronizationManager::GetHadrons(vector<shared_ptr<Hadron>>& signal){
 	//signal = outHadrons;
 	signal.clear();
 	// foreach hadronizon object tasks
-	for(shared_ptr<JetScapeTask> it : GetTaskList()){ 
+	for(shared_ptr<JetScapeTask> it : GetTaskList()){
 		vector<shared_ptr<Hadron>> tempHadronList;
 		JetScapeTask *jet = it.get();
 		Hadronization *hit = (Hadronization *) jet;
@@ -147,6 +147,24 @@ void HadronizationManager::GetHadrons(vector<shared_ptr<Hadron>>& signal){
 		for(auto hadron : tempHadronList){
 			signal.push_back(hadron);
 		}
+	}
+}
+
+void HadronizationManager::DeleteHadrons() {
+  // foreach hadronizon object tasks
+	for(shared_ptr<JetScapeTask> it : GetTaskList()){
+		JetScapeTask *jet = it.get();
+		Hadronization *hit = (Hadronization *) jet;
+		hit->DeleteHadrons();
+	}
+}
+
+void HadronizationManager::DeleteRealHadrons() {
+  // foreach hadronizon object tasks
+	for(shared_ptr<JetScapeTask> it : GetTaskList()){
+		JetScapeTask *jet = it.get();
+		Hadronization *hit = (Hadronization *) jet;
+		hit->DeleteRealHadrons();
 	}
 }
 
