@@ -2,7 +2,7 @@
  * Copyright (c) The JETSCAPE Collaboration, 2018
  *
  * Modular, task-based framework for simulating all aspects of heavy-ion collisions
- * 
+ *
  * For the list of contributors see AUTHORS.
  *
  * Report issues at https://github.com/JETSCAPE/JETSCAPE/issues
@@ -46,6 +46,15 @@ public:
 
 		//get Hadrons from Hadronization submodules
   void GetHadrons(vector<shared_ptr<Hadron>>& signal);
+
+  // deletes the hadrons from the different hadronization modules
+  // this is used in the case of hadronization hadrons in the afterburner
+  // otherwise these hadrons are printed to file and the same hadrons will be
+  // modified in the transport and printed again
+  void DeleteHadrons();
+  // this function removes all positive hadrons, the negative ones are not deleted
+  // needed, when positive hadrons are given to the afterburner
+  void DeleteRealHadrons();
 
   sigslot::signal1<vector<shared_ptr<Hadron>> &> GetHadronList; //get Hadrons from HardProcess NOT Hadronization submodules
 
