@@ -64,12 +64,16 @@ void iSpectraSamplerWrapper::InitTask() {
   iSpectraSampler_ptr_->paraRdr_ptr->readFromFile(input_file);
 
   // overwrite some parameters
+  int echoLevel = GetXMLElementInt({"vlevel"});
+  iSpectraSampler_ptr_->paraRdr_ptr->setVal("JSechoLevel", echoLevel);
+
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("hydro_mode", hydro_mode);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("afterburner_type",
                                             afterburner_type);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("output_samples_into_files", 0);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("use_OSCAR_format", 0);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("use_gzip_format", 0);
+  iSpectraSampler_ptr_->paraRdr_ptr->setVal("use_binary_format", 0);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("store_samples_in_memory", 1);
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("number_of_repeated_sampling",
                                             number_of_repeated_sampling);
@@ -92,7 +96,7 @@ void iSpectraSamplerWrapper::InitTask() {
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("f0_is_not_small", 1);
 
   iSpectraSampler_ptr_->paraRdr_ptr->setVal("calculate_vn", 0);
-  iSpectraSampler_ptr_->paraRdr_ptr->setVal("MC_sampling", 2);
+  iSpectraSampler_ptr_->paraRdr_ptr->setVal("MC_sampling", 4);
 
   iSpectraSampler_ptr_->paraRdr_ptr->setVal(
       "sample_upto_desired_particle_number", 0);
@@ -100,6 +104,8 @@ void iSpectraSamplerWrapper::InitTask() {
 }
 
 void iSpectraSamplerWrapper::Exec() {
+  JSINFO << "running iSS ...";
+
   // generate symbolic links with music_input_file
   std::string music_input_file_path = GetXMLElementText(
           {"Hydro", "MUSIC", "MUSIC_input_file"});
@@ -131,6 +137,7 @@ void iSpectraSamplerWrapper::Exec() {
     exit(-1);
   }
   PassHadronListToJetscape();
+  JSINFO << "iSS finished.";
 }
 
 void iSpectraSamplerWrapper::Clear() {
