@@ -54,6 +54,9 @@ void MpiMusic::InitializeHydro(Parameter parameter_list) {
   music_hydro_ptr = std::unique_ptr<MUSIC>(new MUSIC(input_file));
 
   // overwrite input options
+  int echoLevel = GetXMLElementInt({"vlevel"});
+  music_hydro_ptr->set_parameter("JSechoLevel", echoLevel);
+
   flag_output_evo_to_file = (
       GetXMLElementInt({"Hydro", "MUSIC", "output_evolution_to_file"}));
   if (flag_output_evo_to_file == 1) {
@@ -62,6 +65,8 @@ void MpiMusic::InitializeHydro(Parameter parameter_list) {
   double tau_hydro = (
           GetXMLElementDouble({"Hydro", "MUSIC", "Initial_time_tau_0"}));
   music_hydro_ptr->set_parameter("Initial_time_tau_0", tau_hydro);
+  int beastMode = GetXMLElementInt({"Hydro", "MUSIC", "beastMode"});
+  music_hydro_ptr->set_parameter("beastMode", beastMode);
 
   double eta_over_s =
       GetXMLElementDouble({"Hydro", "MUSIC", "shear_viscosity_eta_over_s"});
@@ -81,23 +86,26 @@ void MpiMusic::InitializeHydro(Parameter parameter_list) {
       GetXMLElementInt({"Hydro", "MUSIC", "T_dependent_Shear_to_S_ratio"}));
   if (flag_shear_Tdep > 0) {
     music_hydro_ptr->set_parameter("Viscosity_Flag_Yes_1_No_0", 1);
+    music_hydro_ptr->set_parameter("T_dependent_Shear_to_S_ratio",
+                                   flag_shear_Tdep);
     if (flag_shear_Tdep == 3) {
       double shear_kinkT = (
         GetXMLElementDouble({"Hydro", "MUSIC", "eta_over_s_T_kink_in_GeV"}));
-      music_hydro_ptr->set_parameter("eta_over_s_T_kink_in_GeV", shear_kinkT);
+      music_hydro_ptr->set_parameter("shear_viscosity_3_T_kink_in_GeV",
+                                     shear_kinkT);
       double shear_lowTslope = (
         GetXMLElementDouble({"Hydro", "MUSIC",
                              "eta_over_s_low_T_slope_in_GeV"}));
-      music_hydro_ptr->set_parameter("eta_over_s_low_T_slope_in_GeV",
+      music_hydro_ptr->set_parameter("shear_viscosity_3_low_T_slope_in_GeV",
                                      shear_lowTslope);
       double shear_highTslope = (
         GetXMLElementDouble({"Hydro", "MUSIC",
                              "eta_over_s_high_T_slope_in_GeV"}));
-      music_hydro_ptr->set_parameter("eta_over_s_high_T_slope_in_GeV",
+      music_hydro_ptr->set_parameter("shear_viscosity_3_high_T_slope_in_GeV",
                                      shear_highTslope);
       double shear_kink = (
         GetXMLElementDouble({"Hydro", "MUSIC", "eta_over_s_at_kink"}));
-      music_hydro_ptr->set_parameter("eta_over_s_at_kink", shear_kink);
+      music_hydro_ptr->set_parameter("shear_viscosity_3_at_kink", shear_kink);
     }
   }
 
@@ -110,17 +118,19 @@ void MpiMusic::InitializeHydro(Parameter parameter_list) {
     if (flag_bulkvis == 3) {
         double bulk_max = GetXMLElementDouble(
               {"Hydro", "MUSIC", "zeta_over_s_max"});
-        music_hydro_ptr->set_parameter("zeta_over_s_max", bulk_max);
+        music_hydro_ptr->set_parameter("bulk_viscosity_3_max", bulk_max);
         double bulk_peakT = GetXMLElementDouble(
               {"Hydro", "MUSIC", "zeta_over_s_T_peak_in_GeV"});
-        music_hydro_ptr->set_parameter("zeta_over_s_T_peak_in_GeV",
+        music_hydro_ptr->set_parameter("bulk_viscosity_3_T_peak_in_GeV",
                                        bulk_peakT);
         double bulk_width = GetXMLElementDouble(
               {"Hydro", "MUSIC", "zeta_over_s_width_in_GeV"});
-        music_hydro_ptr->set_parameter("zeta_over_s_width_in_GeV", bulk_width);
+        music_hydro_ptr->set_parameter("bulk_viscosity_3_width_in_GeV",
+                                       bulk_width);
         double bulk_asy = GetXMLElementDouble(
               {"Hydro", "MUSIC", "zeta_over_s_lambda_asymm"});
-        music_hydro_ptr->set_parameter("zeta_over_s_lambda_asymm", bulk_asy);
+        music_hydro_ptr->set_parameter("bulk_viscosity_3_lambda_asymm",
+                                       bulk_asy);
     }
   }
 
