@@ -15,8 +15,8 @@
 
 // Create a pythia collision at a specified point and return the two inital hard partons
 
-#ifndef PYTHIAGUN_H
-#define PYTHIAGUN_H
+#ifndef EPEMGUN_H
+#define EPEMGUN_H
 
 #include "HardProcess.h"
 #include "JetScapeLogger.h"
@@ -24,44 +24,47 @@
 
 using namespace Jetscape;
 
-class PythiaGun : public HardProcess, public Pythia8::Pythia {
+class epemGun : public HardProcess, public Pythia8::Pythia {
 
 private:
-  double pTHatMin;
-  double pTHatMax;
+  //double pTHatMin;
+  //double pTHatMax;
   double eCM;
-  double vir_factor;
-  double softMomentumCutoff;
-  bool FSR_on;
-  bool softQCD;
+  //bool FSR_on;
 
   // Allows the registration of the module so that it is available to be used by the Jetscape framework.
-  static RegisterJetScapeModule<PythiaGun> reg;
+  static RegisterJetScapeModule<epemGun> reg;
+
+  double sud_val_QG_w_M(double M, double h0, double h1, double h2, double E1);
+  double sud_z_QG_w_M(double M, double cg, double cg1, double E2);
+  double alpha_s(double q2);
+
+protected:
+  std::uniform_real_distribution<double> ZeroOneDistribution;
 
 public:
   /** standard ctor
       @param xmlDir: Note that the environment variable PYTHIA8DATA takes precedence! So don't use it.
       @param printBanner: Suppress starting blurb. Should be set to true in production, credit where it's due
   */
-  PythiaGun(string xmlDir = "DONTUSETHIS", bool printBanner = false)
+  epemGun(string xmlDir = "DONTUSETHIS", bool printBanner = false)
       : Pythia8::Pythia(xmlDir, printBanner), HardProcess() {
-    SetId("UninitializedPythiaGun");
+    SetId("UninitializedepemGun");
   }
 
-  ~PythiaGun();
+  ~epemGun();
 
   void InitTask();
   void Exec();
 
   // Getters
-  double GetpTHatMin() const { return pTHatMin; }
-  double GetpTHatMax() const { return pTHatMax; }
+  //double GetpTHatMin() const { return pTHatMin; }
+  //double GetpTHatMax() const { return pTHatMax; }
 
   // Cross-section information in mb and event weight.
   double GetSigmaGen() { return info.sigmaGen(); };
   double GetSigmaErr() { return info.sigmaErr(); };
-  double GetPtHat() { return info.pTHat(); };
   double GetEventWeight() { return info.weight(); };
 };
 
-#endif // PYTHIAGUN_H
+#endif // EPEMGUN_H
